Develop - 30/03/2023

+++ src/main/java/com/oscarjrod/eurderapi/items/domain/ItemDto.java
@@ -1,11 +1,13 @@
 package com.oscarjrod.eurderapi.items.domain;
 
+import java.math.BigDecimal;
+
 public class ItemDto {
 
     private Long id;
     private String name;
     private String description;
-    private double price;
+    private BigDecimal price;
     private long stock;
 
     public static ItemDto createItemDto(Item item) {
@@ -42,11 +44,11 @@ public class ItemDto {
         this.description = description;
     }
 
-    public double getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 
-    public void setPrice(double price) {
+    public void setPrice(BigDecimal price) {
         this.price = price;
     }
